Separating connection specifics into its own set of extensions
