+++ Source/DotNET/EntityFrameworkCore/ReadOnlyDbContextExtensions.cs
@@ -51,26 +51,6 @@ public static class ReadOnlyDbContextExtensions
     }
 
     /// <summary>
-    /// Configures the DbContext to use the database specified in the connection string.
-    /// The database type is inferred from the connection string.
-    /// </summary>
-    /// <param name="builder">The DbContext options builder to configure.</param>
-    /// <param name="connectionString">The connection string to use.</param>
-    /// <returns>The configured DbContext options builder.</returns>
-    /// <exception cref="UnsupportedDatabaseType">Thrown if the connection string does not have a supported database type.</exception>
-    public static DbContextOptionsBuilder UseDatabaseFromConnectionString(this DbContextOptionsBuilder builder, string connectionString)
-    {
-        var type = connectionString.GetDatabaseType();
-        return type switch
-        {
-            DatabaseType.Sqlite => builder.UseSqlite(connectionString),
-            DatabaseType.SqlServer => builder.UseSqlServer(connectionString),
-            DatabaseType.PostgreSql => builder.UseNpgsql(connectionString),
-            _ => throw new UnsupportedDatabaseType(connectionString)
-        };
-    }
-
-    /// <summary>
     /// Adds a read-only DbContext to the service collection, configured with the provided connection string
     /// and database type.
     /// </summary>
@@ -90,18 +70,4 @@ public static class ReadOnlyDbContextExtensions
 
         return services;
     }
-
-    /// <summary>
-    /// Infers the database type from the connection string.
-    /// </summary>
-    /// <param name="connectionString">The connection string to infer the database type from.</param>
-    /// <returns>The inferred database type.</returns>
-    /// <exception cref="UnsupportedDatabaseType">Thrown if the connection string does not have a supported database type.</exception>
-    public static DatabaseType GetDatabaseType(this string connectionString) => connectionString switch
-    {
-        _ when connectionString.Contains("Data Source=") || connectionString.Contains("Filename=") => DatabaseType.Sqlite,
-        _ when connectionString.Contains("Server=") && connectionString.Contains("Database=") => DatabaseType.SqlServer,
-        _ when connectionString.Contains("Host=") && connectionString.Contains("Database=") => DatabaseType.PostgreSql,
-        _ => throw new UnsupportedDatabaseType(connectionString)
-    };
 }
